requirements.txt
streamlit==1.16.0
gspread==5.7.2
numpy==1.23.5
python-docx==0.8.11
altair==4.0
